secure_file_sharing/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-multipart
pyjwt
email-validator
aiofiles